+++ examples/person.sql
@@ -20,8 +20,8 @@ create table person (
   additional_name_as_en varchar, -- example: Amy
   additional_name_as_es varchar, -- example: Amé
   additional_name_as_fr varchar, -- example: Aimée
-  height float, -- example: 1.65 meters
-  mass float, -- example: 70000 grams 
+  height decimal(20,12), -- example: 1.65 meters
+  mass decimal(20,12), -- example: 70000 grams 
   geolocation_id uuid references geolocation, -- example: where the person is now
   global_location_number decimal(13), -- https://en.wikipedia.org/wiki/Global_Location_Number
   lifespan_id uuid references lifespan -- example: birth date is 2000-01-01
